requirements.txt
Flask==2.1.2
pycryptodome==3.11.0
